feat!(artifact-caching-proxy) introduces repository fallbacks chain
The goal is to improve build performances by:

- Caching Maven central artifacts in ACP along with Jenkins artefacts
- Delegate the chain of "retries" to Nginx (instead of Maven) for efficient caching
- Tune the network behavior by disabling IPv6 DNS resolution and using default Nginx timeouts

The fallback setup is defined statically in the ConfigMap. Possible improvement would be to templatize the chain of fallbacks.
For this version the chain is the following: "Jenkins public repo" -> "Jenkins incrementals repo" -> "Maven Central repo".

Notes:

- The number of possible fallbacks is limited internally in Nginx to 10 (ref. http://nginx.org/en/docs/http/ngx_http_core_module.html#recursive_error_pages)
- The non-cached requests (maven-metadata.xml files) are also chained on their own using the same (Possible improvement by templatizing)

BREAKING CHANGE: For each 'mirror.url' in the `~/.m2/settings.xml` pointing to ACP, there should be no URI (path) otherwise ACP will answer HTTP/404.
  Example: '<url>http://localhost:8888/</url>' instead of '<url>http://localhost:8888/public/</url>'.

Signed-off-by: Damien Duportal <damien.duportal@gmail.com>